+++ pages/launch.tsx
@@ -1,31 +1,14 @@
 import * as React from 'react';
-import Image from 'next/image';
 import { useRouter } from 'next/router';
 
-import { Alert, Box, Button, CircularProgress, Typography } from '@mui/joy';
+import { Alert, Box, Button, Typography } from '@mui/joy';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 
 import { useComposerStore } from '../src/apps/chat/components/composer/store-composer';
 
 import { AppLayout } from '~/common/layout/AppLayout';
+import { LogoProgress } from '~/common/components/LogoProgress';
 import { asValidURL } from '~/common/util/urlUtils';
-
-
-export function LogoProgress(props: { showProgress: boolean }) {
-  return <Box sx={{
-    width: 64,
-    height: 64,
-    position: 'relative',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  }}>
-    <Box sx={{ position: 'absolute', mt: 0.75 }}>
-      <Image src='/icons/favicon-32x32.png' alt='App Logo' width={32} height={32} />
-    </Box>
-    {props.showProgress && <CircularProgress size='lg' sx={{ position: 'absolute' }} />}
-  </Box>;
-}
 
 
 /**
